refactor: update import paths and arktype port validation in template generation
